requirements.txt
streamlit
google-generativeai
duckduckgo-search
pypdf
python-docx
beautifulsoup4
requests